requirements.txt
ipykernel
pandas
numpy
scikit-learn
pdfplumber
pdfminer
openpyxl
google-generativeai
Pillow
PyPDF2
pdf2image
python-dotenv
streamlit